+++ crates/plotinator-supported-formats/src/hdf5.rs
@@ -1,5 +1,6 @@
 use chrono::{DateTime, Utc};
 use plotinator_hdf5::frame_altimeters::FrameAltimeters;
+use plotinator_hdf5::frame_inclinometers::FrameInclinometers;
 use plotinator_hdf5::njord_ins::NjordIns;
 use plotinator_hdf5::{bifrost::BifrostLoopCurrent, wasp200::Wasp200};
 use plotinator_log_if::prelude::*;
@@ -82,5 +83,6 @@ define_supported_hdf5_formats! {
     BifrostLoopCurrent => BifrostLoopCurrent,
     Wasp200Height => Wasp200,
     FrameAltimeters => FrameAltimeters,
+    FrameInclinometers => FrameInclinometers,
     NjordIns => NjordIns
 }