+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Support for `Frame Inclinometers` HDF5 files
+
 ## [1.20.2]
 
 ### Changed